+++ lib/ecto/adapters/clickhouse.ex
@@ -72,41 +72,6 @@ defmodule Ecto.Adapters.ClickHouse do
       def disconnect_all(interval, opts \\ []) do
         Ecto.Adapters.SQL.disconnect_all(get_dynamic_repo(), interval, opts)
       end
-
-      @doc """
-      Similar to `insert_all/2` but with the following differences:
-
-        - accepts rows as streams or lists
-        - sends rows as a chunked request
-        - doesn't autogenerate ids or does any other preprocessing
-
-      Example:
-
-          Repo.query!("create table ecto_ch_demo(a UInt64, b String) engine Null")
-
-          defmodule Demo do
-            use Ecto.Schema
-
-            @primary_key false
-            schema "ecto_ch_demo" do
-              field :a, Ch, type: "UInt64"
-              field :b, :string
-            end
-          end
-
-          rows = Stream.map(1..100_000, fn i -> %{a: i, b: to_string(i)} end)
-          {100_000, nil} = Repo.insert_stream(Demo, rows)
-
-          # schemaless
-          {100_000, nil} = Repo.insert_stream("ecto_ch_demo", rows, types: [a: Ch.Types.u64(), b: :string])
-
-      """
-      def insert_stream(source_or_schema, rows, opts \\ []) do
-        repo = get_dynamic_repo()
-        # TODO need it?
-        # opts = Ecto.Repo.Supervisor.tuplet(repo, prepare_opts(:insert_all, opts))
-        Ecto.Adapters.ClickHouse.Schema.insert_stream(repo, source_or_schema, rows, opts)
-      end
     end
   end
 
@@ -242,13 +207,18 @@ defmodule Ecto.Adapters.ClickHouse do
       on_conflict,
       returning,
       placeholders,
-      opts
+      [{:command, :insert} | opts]
     )
   end
 
   @impl Ecto.Adapter.Schema
   def insert(adapter_meta, schema_meta, params, _, _, opts) do
-    Ecto.Adapters.ClickHouse.Schema.insert(adapter_meta, schema_meta, params, opts)
+    Ecto.Adapters.ClickHouse.Schema.insert(
+      adapter_meta,
+      schema_meta,
+      params,
+      [{:command, :insert} | opts]
+    )
   end
 
   @dialyzer {:no_return, update: 6}
@@ -277,12 +247,22 @@ defmodule Ecto.Adapters.ClickHouse do
   # https://github.com/elixir-ecto/ecto/pull/4277
   # @impl Ecto.Adapter.Schema
   def delete(adapter_meta, schema_meta, params, opts) do
-    Ecto.Adapters.ClickHouse.Schema.delete(adapter_meta, schema_meta, params, opts)
+    Ecto.Adapters.ClickHouse.Schema.delete(
+      adapter_meta,
+      schema_meta,
+      params,
+      [{:command, :delete} | opts]
+    )
   end
 
   @impl Ecto.Adapter.Schema
   def delete(adapter_meta, schema_meta, params, _returning, opts) do
-    Ecto.Adapters.ClickHouse.Schema.delete(adapter_meta, schema_meta, params, opts)
+    Ecto.Adapters.ClickHouse.Schema.delete(
+      adapter_meta,
+      schema_meta,
+      params,
+      [{:command, :delete} | opts]
+    )
   end
 
   @impl Ecto.Adapter.Queryable